requirements.txt
numpy
matplotlib
imageio==2.*
Pillow
scikit-learn
opencv-python
scikit-surprise